Chore: move auxiliary functions to utilities
